+++ terraform/modules/s3-bucket/variables.tf
@@ -1,0 +1,28 @@
+variable "bucket_name" {
+  description = "Name of the S3 bucket"
+  type        = string
+}
+
+variable "enable_versioning" {
+  description = "Enable bucket versioning"
+  type        = bool
+  default     = true
+}
+
+variable "enable_encryption" {
+  description = "Enable server-side encryption"
+  type        = bool
+  default     = true
+}
+
+variable "enable_access_logging" {
+  description = "Enable S3 access logging"
+  type        = bool
+  default     = false
+}
+
+variable "tags" {
+  description = "Tags to apply to resources"
+  type        = map(string)
+  default     = {}
+}